requirements.txt
# Core dependencies
mcp>=1.9.0
pytz>=2023.3
python-dotenv>=1.0.0
httpx>=0.24.0

# Data processing
python-dateutil>=2.8.0

# Optional visualization dependencies
plotly>=5.17.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0